+++ Rangefinder.h
@@ -21,9 +21,9 @@ public:
 	void constrain_params(void);
 	Parameters *params;
 
-	uint32_t PINGRATE;
-	uint32_t RANGE_ENABLE;
-	uint32_t LPF_ENABLE;
+	uint8_t PINGRATE;
+	uint8_t RANGE_ENABLE;
+	uint8_t LPF_ENABLE;
 	float LPF_CUTOFF;
 
 	//TODO parse and store as uint16_t
